Adicionando end-points de inserir e listar clientes, com imports ajustados.

+++ Projeto-Bank-Accenture/src/main/java/com/bankaccenture/Projeto_Bank_Accenture/model/Agencia.java
@@ -25,7 +25,7 @@ import lombok.NoArgsConstructor;
 public class Agencia {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long idAgencia;
+	private int idAgencia;
 	
 	@Column
 	private StringBuffer nomeAgencia;
