+++ Projeto-Bank-Accenture/src/main/java/com/bankaccenture/Projeto_Bank_Accenture/repository/AgenciaRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.bankaccenture.Projeto_Bank_Accenture.model.Agencia;
 
-public interface AgenciaRepository extends JpaRepository<Agencia,Long>{
+public interface AgenciaRepository extends JpaRepository<Agencia,Integer>{
 
 }
