+++ Projeto-Bank-Accenture/src/main/java/com/bankaccenture/Projeto_Bank_Accenture/repository/ContaCorrenteRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.bankaccenture.Projeto_Bank_Accenture.model.ContaCorrente;
 
-public interface ContaCorrenteRepository extends JpaRepository<ContaCorrente,Long>{
+public interface ContaCorrenteRepository extends JpaRepository<ContaCorrente,Integer>{
 
 }
