+++ Projeto-Bank-Accenture/src/main/java/com/bankaccenture/Projeto_Bank_Accenture/repository/ExtratoRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.bankaccenture.Projeto_Bank_Accenture.model.Extrato;
 
-public interface ExtratoRepository extends JpaRepository<Extrato,Long>{
+public interface ExtratoRepository extends JpaRepository<Extrato,Integer>{
 
 }
